requirements.txt
mpu6050==0.0.3
numpy==2.3.1
piper_tts==1.2.0
playsound==1.3.0
rapidfuzz==3.13.0
sounddevice==0.5.1
whispercpp==0.0.17
xpinyin==0.7.6
